Add a special project-level metadata property called "server", with a subproperty called "staticPaths".
Set server.staticPaths to an array of directory paths (relative to the
project directory) that you want to serve static files from when running
Selleck with the --server option.

This is used in YUI to serve dev builds of the YUI source during
documentation/example development (production builds are used at all
other times).